typing fixes and cleanup

+++ cilantro_ee/contracts/genesis/currency.s.py
@@ -5,7 +5,7 @@ def seed(vk: str):
     balances[vk] = 288_090_567
 
 @export
-def transfer(amount: int, to: str):
+def transfer(amount: float, to: str):
     sender = ctx.caller
     assert balances[sender] >= amount, 'Not enough coins to send!'
 
@@ -25,13 +25,13 @@ def allowance(owner: str, spender: str):
     return balances[owner, spender]
 
 @export
-def approve(amount: int, to: str):
+def approve(amount: float, to: str):
     sender = ctx.caller
     balances[sender, to] += amount
     return balances[sender, to]
 
 @export
-def transfer_from(amount: int, to: str, main_account: str):
+def transfer_from(amount: float, to: str, main_account: str):
     sender = ctx.caller
 
     assert balances[main_account, sender] >= amount, 'Not enough coins approved to send! You have {} and are trying to spend {}'\
